% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv.obj.R
\name{adv.obj}
\alias{adv.obj}
\title{Advance individuals using internal functions}
\usage{
adv.obj(obj, times = 1)
}
\arguments{
\item{times}{}
}
\value{
list object
}
\description{
Advance individuals using internal functions
}
\examples{
obj <- create.obj(n=10)
obj <- adv.obj(obj, times = 5)
obj

}
