% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.obj.R
\name{create.obj}
\alias{create.obj}
\title{Create object}
\usage{
create.obj(n = 10)
}
\arguments{
\item{n}{number ind individuals}
}
\value{
list object
}
\description{
Create object
}
\examples{
obj <- create.obj(n = 20)
obj

}
